moving simple output handler into handlers/output
