+++ src/handlers/mod.rs
@@ -8,7 +8,7 @@ pub mod outputs;
 // export Handlers
 pub use debug::DebugHandler;
 pub use msg_count::MessageCountHandler;
-pub use outputs::OutputHandler;
+pub use outputs::SimpleOutputHandler;
 
 #[async_trait::async_trait]
 pub trait Handler: Debug + Send + Sync {
